added a check for signed messages begin and end lines
The regex would check the correct formatting of the pgp messages, it would search for the existence of certain phrases for this functionality, but because the regex was implemented in a multiline method it could not see if these phrases were added multiple times, which would be incorrect.
For this we added some extra regex checks to make sure these phrases are not added multiple times.